refactor(utils): extract llm tests
Amp-Thread-ID: https://ampcode.com/threads/T-019dc83e-76e3-7306-a2f3-8c6e105a3dae
